fix: Fast packet assembly (#39)
The available information is ambiguous on whether fast packet frames
must arrive in order. The previous implementation allowed out of order
arrival, including allowing subsequent frames to precede the first frame
(which includes the length of the complete fast frame). As a result
missing a frame might create a cascade of failures in receiving a fast
packet from that source and pgn using the same sequence ID.

The changes tighten the ordering requirement: the first frame must
precede all others. This assures that if we miss any frame in the
sequence (possible since the transport does not guarantee delivery) the
previous (partial) fast packet will be reset and we'll resync with the
sender the next time they send a sequence for the PGN with the same
sequence ID.

---------

Co-authored-by: Russell Daniels <russd@inquest.com>